decouple LAN logic from UI for network_game
